+++ ext/standard/head.c
@@ -180,7 +180,7 @@ PHP_FUNCTION(setcookie)
 	char *name, *value = NULL, *path = NULL, *domain = NULL;
 	zend_long expires = 0;
 	zend_bool secure = 0, httponly = 0;
-	int name_len, value_len = 0, path_len = 0, domain_len = 0;
+	size_t name_len, value_len = 0, path_len = 0, domain_len = 0;
 
 	if (zend_parse_parameters(ZEND_NUM_ARGS() TSRMLS_CC, "s|slssbb", &name,
 							  &name_len, &value, &value_len, &expires, &path,
@@ -203,7 +203,7 @@ PHP_FUNCTION(setrawcookie)
 	char *name, *value = NULL, *path = NULL, *domain = NULL;
 	zend_long expires = 0;
 	zend_bool secure = 0, httponly = 0;
-	int name_len, value_len = 0, path_len = 0, domain_len = 0;
+	size_t name_len, value_len = 0, path_len = 0, domain_len = 0;
 
 	if (zend_parse_parameters(ZEND_NUM_ARGS() TSRMLS_CC, "s|slssbb", &name,
 							  &name_len, &value, &value_len, &expires, &path,
